Back/chore/gnidinger/oauth2

+++ back/src/main/java/seb40_main_012/back/pairing/PairingService.java
@@ -12,7 +12,9 @@ import seb40_main_012.back.book.BookService;
 import seb40_main_012.back.book.bookInfoSearchAPI.BookInfoSearchService;
 import seb40_main_012.back.book.entity.Book;
 import seb40_main_012.back.common.bookmark.BookmarkRepository;
+import seb40_main_012.back.common.comment.CommentRepository;
 import seb40_main_012.back.common.comment.entity.Comment;
+import seb40_main_012.back.common.comment.entity.CommentType;
 import seb40_main_012.back.common.like.LikeRepository;
 import seb40_main_012.back.pairing.entity.Pairing;
 import seb40_main_012.back.user.entity.User;
@@ -37,6 +39,7 @@ public class PairingService {
     private final UserService userService;
     private final UserRepository userRepository;
     private final BookInfoSearchService bookInfoSearchService;
+    private final CommentRepository commentRepository;
 
     public Pairing createPairing(Pairing pairing, String isbn13) {
 
@@ -68,26 +71,29 @@ public class PairingService {
 
         Pairing findPairing = findVerifiedPairing(pairingId);
 
-        Pairing updatedPairing =
-                Pairing.builder()
-                        .book(findPairing.getBook())
-                        .user(findPairing.getUser())
-                        .pairingId(findPairing.getPairingId())
-                        .imagePath(pairing.getImagePath())
-                        .title(pairing.getTitle())
-                        .body(pairing.getBody())
-                        .pairingCategory(pairing.getPairingCategory())
-                        .outLinkPath(pairing.getOutLinkPath())
-                        .likeCount(findPairing.getLikeCount())
-                        .view(findPairing.getView())
-                        .image(findPairing.getImage())
-                        .comments(findPairing.getComments())
-                        .likes(findPairing.getLikes())
-                        .createdAt(findPairing.getCreatedAt())
-                        .modifiedAt(LocalDateTime.now())
-                        .build();
+        if (findUser == findPairing.getUser()) {
 
-        return pairingRepository.save(updatedPairing);
+            Pairing updatedPairing =
+                    Pairing.builder()
+                            .book(findPairing.getBook())
+                            .user(findPairing.getUser())
+                            .pairingId(findPairing.getPairingId())
+                            .imagePath(pairing.getImagePath())
+                            .title(pairing.getTitle())
+                            .body(pairing.getBody())
+                            .pairingCategory(pairing.getPairingCategory())
+                            .outLinkPath(pairing.getOutLinkPath())
+                            .likeCount(findPairing.getLikeCount())
+                            .view(findPairing.getView())
+                            .image(findPairing.getImage())
+                            .comments(findPairing.getComments())
+                            .likes(findPairing.getLikes())
+                            .createdAt(findPairing.getCreatedAt())
+                            .modifiedAt(LocalDateTime.now())
+                            .build();
+
+            return pairingRepository.save(updatedPairing);
+        } else throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED);
     }
 
     public Pairing addLike(long pairingId) {
@@ -180,7 +186,7 @@ public class PairingService {
 
         Boolean isBookmarked;
 
-        if (bookmarkRepository.findByUserAndPairing(findUser,pairing) == null) { //북마크 안 누른 경우
+        if (bookmarkRepository.findByUserAndPairing(findUser, pairing) == null) { //북마크 안 누른 경우
             isBookmarked = false;
         } else {
             isBookmarked = true;
@@ -196,7 +202,7 @@ public class PairingService {
         return findVerifiedPairing(pairingId);
     }
 
-//    --------------------------------------------------------------------------------------------
+    //    --------------------------------------------------------------------------------------------
 //    --------------------------------------------------------------------------------------------
 //    조회 API 세분화
 //    --------------------------------------------------------------------------------------------
@@ -229,6 +235,13 @@ public class PairingService {
         return pairingRepository.findCategorySliceByNewestDesc("FILM", pageRequest);
     }
 
+    public List<Pairing> findFilmPairingsRandom() { // 영화 카테고리 무작위
+
+        PageRequest pageRequest = PageRequest.of(0, 10);
+
+        return pairingRepository.findAllByCategory("FILM", pageRequest);
+    }
+
     public List<Pairing> findCuisinePairingsLikes() { // 음식/장소 카테고리 슬라이스 처리 및 좋아요 내림차순 정렬
 
         PageRequest pageRequest = PageRequest.of(0, 10);
@@ -242,6 +255,14 @@ public class PairingService {
 
 
         return pairingRepository.findCategorySliceByNewestDesc("CUISINE", pageRequest);
+    }
+
+    public List<Pairing> findCuisinePairingsRandom() { // 음식/장소 카테고리 무작위
+
+        PageRequest pageRequest = PageRequest.of(0, 10);
+
+
+        return pairingRepository.findAllByCategory("CUISINE", pageRequest);
     }
 
     public List<Pairing> findMusicPairingsLikes() { // 음악 카테고리 슬라이스 처리 및 좋아요 내림차순 정렬
@@ -259,6 +280,14 @@ public class PairingService {
         return pairingRepository.findCategorySliceByNewestDesc("MUSIC", pageRequest);
     }
 
+    public List<Pairing> findMusicPairingsRandom() { // 음악 카테고리 무작위
+
+        PageRequest pageRequest = PageRequest.of(0, 10);
+
+
+        return pairingRepository.findAllByCategory("MUSIC", pageRequest);
+    }
+
     public List<Pairing> findBookPairingsLikes() { // 책 카테고리 슬라이스 처리 및 좋아요 내림차순 정렬
 
         PageRequest pageRequest = PageRequest.of(0, 10);
@@ -270,8 +299,14 @@ public class PairingService {
 
         PageRequest pageRequest = PageRequest.of(0, 10, Sort.by(Sort.Direction.DESC, "created_At"));
 
-
         return pairingRepository.findCategorySliceByNewestDesc("BOOK", pageRequest);
+    }
+
+    public List<Pairing> findBookPairingsRandom() { // 책 카테고리 무작위
+
+        PageRequest pageRequest = PageRequest.of(0, 10);
+
+        return pairingRepository.findAllByCategory("BOOK", pageRequest);
     }
 
     public List<Pairing> findEtcPairingsLikes() { // 기타 카테고리 슬라이스 처리 및 좋아요 내림차순 정렬
@@ -288,9 +323,21 @@ public class PairingService {
         return pairingRepository.findCategorySliceByNewestDesc("ETC", pageRequest);
     }
 
-    public List<Pairing> findBestPairingsLikes() { // 기타 카테고리 슬라이스 처리 및 등록일 내림차순 정렬
+    public List<Pairing> findEtcPairingsRandom() { // 기타 카테고리 무작위
+
+        PageRequest pageRequest = PageRequest.of(0, 10);
+
+        return pairingRepository.findAllByCategory("ETC", pageRequest);
+    }
+
+    public List<Pairing> findBestPairingsLikes() {
 
         return pairingRepository.findBestTenCategory();
+    }
+
+    public List<Pairing> findRandomPairings() {
+
+        return pairingRepository.findAll();
     }
 
 //    --------------------------------------------------------------------------------------------
@@ -312,7 +359,23 @@ public class PairingService {
 
         Pairing findPairing = findVerifiedPairing(pairingId);
 
-        pairingRepository.delete(findPairing);
+        if (findUser == findPairing.getUser()) {
+            commentRepository.deleteAllByPairingId(pairingId);
+            pairingRepository.delete(findPairing);
+        } else throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED);
+    }
+
+    public void deletePairings() {
+
+        User findUser = userService.getLoginUser();
+
+        long userId = findUser.getUserId();
+
+        findUser.getPairings().stream()
+                .map(Pairing::getPairingId)
+                .forEach(commentRepository::deleteAllByPairingId);
+
+        pairingRepository.deleteAllByUserId(userId);
     }
